pyproject.toml
[tool.poetry]
name = "drf-tracker"
version = "0.1.0"
description = ""
authors = ["PatakIN <git@patakin.space>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^24.1"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
